+++ src/generator/writers/writeRouter.ts
@@ -8,20 +8,27 @@ import { resolveFuncType } from "./resolveFuncType";
 import { resolveFuncTypeUnpack } from "./resolveFuncTypeUnpack";
 import { writeStatement } from "./writeFunction";
 
-export function writeRouter(type: TypeDescription, ctx: WriterContext) {
-    ctx.append(`(${resolveFuncType(type, ctx)}, int) ${ops.contractRouter(type.name)}(${resolveFuncType(type, ctx)} self, int msg_bounced, slice in_msg) impure inline_ref {`);
+export function writeRouter(type: TypeDescription, kind: 'internal' | 'external', ctx: WriterContext) {
+    const internal = kind === 'internal';
+    if (internal) {
+        ctx.append(`(${resolveFuncType(type, ctx)}, int) ${ops.contractRouter(type.name, kind)}(${resolveFuncType(type, ctx)} self, int msg_bounced, slice in_msg) impure inline_ref {`);
+    } else {
+        ctx.append(`(${resolveFuncType(type, ctx)}, int) ${ops.contractRouter(type.name, kind)}(${resolveFuncType(type, ctx)} self, slice in_msg) impure inline_ref {`);
+    }
     ctx.inIndent(() => {
 
         // Parse incoming message
         ctx.append();
-
-        ctx.append(`if (msg_bounced) {`);
-        ctx.inIndent(() => {
-            ctx.append(`;; Skip 0xFFFFFFFF`);
-            ctx.append(`in_msg~skip_bits(32);`);
-        });
-        ctx.append(`}`);
-        ctx.append();
+        
+        if (internal) {
+            ctx.append(`if (msg_bounced) {`);
+            ctx.inIndent(() => {
+                ctx.append(`;; Skip 0xFFFFFFFF`);
+                ctx.append(`in_msg~skip_bits(32);`);
+            });
+            ctx.append(`}`);
+            ctx.append();
+        }
 
         ctx.append(`;; Parse incoming message`);
         ctx.append(`int op = 0;`);
@@ -33,75 +40,77 @@ export function writeRouter(type: TypeDescription, ctx: WriterContext) {
         ctx.append(`}`);
         ctx.append();
 
-        // Handle bounced
-        ctx.append(`;; Handle bounced messages`);
-        ctx.append(`if (msg_bounced) {`);
-        ctx.inIndent(() => {
+        if (internal) {
+            // Handle bounced
+            ctx.append(`;; Handle bounced messages`);
+            ctx.append(`if (msg_bounced) {`);
+            ctx.inIndent(() => {
 
-            const nonGenericReceivers = type.receivers.filter(r => {
-                if (r.selector.kind !== "internal-bounce") return false;
-                const allocation = getType(ctx.ctx, r.selector.type);
-                return !(allocation.origin === "stdlib" && allocation.name === "Slice");
-              });
-              
-            const genericReceiver = type.receivers.find(r => {
-                if (r.selector.kind !== "internal-bounce") return false;
-                const allocation = getType(ctx.ctx, r.selector.type);
-                return allocation.origin === "stdlib" && allocation.name === "Slice";
-            });
-            
-            for (const r of nonGenericReceivers) {
-                const selector = r.selector;
-                if (selector.kind !== "internal-bounce") throw Error('Invalid selector type: ' + selector.kind);
-
-                let allocation = getType(ctx.ctx, selector.type);
+                const nonGenericReceivers = type.receivers.filter(r => {
+                    if (r.selector.kind !== "internal-bounce") return false;
+                    const allocation = getType(ctx.ctx, r.selector.type);
+                    return !(allocation.origin === "stdlib" && allocation.name === "Slice");
+                });
                 
-                if (!allocation.header) {
-                    throw Error('Invalid allocation: ' + selector.type);
+                const genericReceiver = type.receivers.find(r => {
+                    if (r.selector.kind !== "internal-bounce") return false;
+                    const allocation = getType(ctx.ctx, r.selector.type);
+                    return allocation.origin === "stdlib" && allocation.name === "Slice";
+                });
+                
+                for (const r of nonGenericReceivers) {
+                    const selector = r.selector;
+                    if (selector.kind !== "internal-bounce") throw Error('Invalid selector type: ' + selector.kind);
+
+                    let allocation = getType(ctx.ctx, selector.type);
+                    
+                    if (!allocation.header) {
+                        throw Error('Invalid allocation: ' + selector.type);
+                    }
+
+                    ctx.append();
+                    ctx.append(`;; Bounced handler for ${selector.type} message`);
+                    ctx.append(`if (op == ${allocation.header}) {`);
+                    ctx.inIndent(() => {
+                        // Read message
+                        ctx.append(`var msg = in_msg~${ops.readerBounced(selector.type, ctx)}();`);
+
+                        // Execute function
+                        ctx.append(`self~${ops.receiveTypeBounce(type.name, selector.type)}(msg);`);
+
+                        // Exit
+                        ctx.append('return (self, true);');
+                    })
+                    ctx.append(`}`);
                 }
 
-                ctx.append();
-                ctx.append(`;; Bounced handler for ${selector.type} message`);
-                ctx.append(`if (op == ${allocation.header}) {`);
-                ctx.inIndent(() => {
-                    // Read message
-                    ctx.append(`var msg = in_msg~${ops.readerBounced(selector.type, ctx)}();`);
+                if (genericReceiver) {
+                    const selector = genericReceiver.selector;
+                    if (selector.kind !== "internal-bounce") throw Error('Invalid selector type: ' + selector.kind);
+
+                    ctx.append();
+                    ctx.append(`;; Bounced handler for ${selector.type} message (Generic)`);
 
                     // Execute function
-                    ctx.append(`self~${ops.receiveTypeBounce(type.name, selector.type)}(msg);`);
+                    ctx.append(`self~${ops.receiveTypeBounce(type.name, selector.type)}(in_msg);`);
 
                     // Exit
                     ctx.append('return (self, true);');
-                })
-                ctx.append(`}`);
-            }
+                } else {
+                    ctx.append(`return (self, true);`);
 
-            if (genericReceiver) {
-                const selector = genericReceiver.selector;
-                if (selector.kind !== "internal-bounce") throw Error('Invalid selector type: ' + selector.kind);
-
-                ctx.append();
-                ctx.append(`;; Bounced handler for ${selector.type} message (Generic)`);
-
-                // Execute function
-                ctx.append(`self~${ops.receiveTypeBounce(type.name, selector.type)}(in_msg);`);
-
-                // Exit
-                ctx.append('return (self, true);');
-            } else {
-                ctx.append(`return (self, true);`);
-
-            }
-            
-        });
-        ctx.append(`}`);
+                }
+                
+            });
+            ctx.append(`}`);
+        }
 
         // Non-empty receivers
         for (const f of type.receivers) {
             const selector = f.selector;
 
             // Generic receiver
-            if (selector.kind === 'internal-binary') {
+            if (selector.kind === (internal ? 'internal-binary' : 'external-binary')) {
                 let allocation = getType(ctx.ctx, selector.type);
                 if (!allocation.header) {
                     throw Error('Invalid allocation: ' + selector.type);
@@ -115,7 +124,7 @@ export function writeRouter(type: TypeDescription, ctx: WriterContext) {
                     ctx.append(`var msg = in_msg~${ops.reader(selector.type, ctx)}();`);
 
                     // Execute function
-                    ctx.append(`self~${ops.receiveType(type.name, selector.type)}(msg);`);
+                    ctx.append(`self~${ops.receiveType(type.name, kind, selector.type)}(msg);`);
 
                     // Exit
                     ctx.append('return (self, true);');
@@ -123,14 +132,14 @@ export function writeRouter(type: TypeDescription, ctx: WriterContext) {
                 ctx.append(`}`);
             }
 
-            if (selector.kind === 'internal-empty') {
+            if (selector.kind === (internal ? 'internal-empty' : 'external-empty')) {
                 ctx.append();
                 ctx.append(`;; Receive empty message`);
                 ctx.append(`if ((op == 0) & (slice_bits(in_msg) <= 32)) {`);
                 ctx.inIndent(() => {
 
                     // Execute function
-                    ctx.append(`self~${ops.receiveEmpty(type.name)}();`);
+                    ctx.append(`self~${ops.receiveEmpty(type.name, kind)}();`);
 
                     // Exit
                     ctx.append('return (self, true);');
@@ -140,17 +149,17 @@ export function writeRouter(type: TypeDescription, ctx: WriterContext) {
         }
 
         // Text resolvers
-        let hasComments = !!type.receivers.find((v) => v.selector.kind === 'internal-comment' || v.selector.kind === 'internal-comment-fallback');
+        let hasComments = !!type.receivers.find((v) => internal ? (v.selector.kind === 'internal-comment' || v.selector.kind === 'internal-comment-fallback') : (v.selector.kind === 'external-comment' || v.selector.kind === 'external-comment-fallback'));
         if (hasComments) {
             ctx.append();
             ctx.append(`;; Text Receivers`);
             ctx.append(`if (op == 0) {`);
             ctx.inIndent(() => {
-                if (!!type.receivers.find((v) => v.selector.kind === 'internal-comment')) {
+                if (!!type.receivers.find((v) => v.selector.kind === (internal ? 'internal-comment' : 'external-comment'))) {
                     ctx.append(`var text_op = slice_hash(in_msg);`);
                     for (const r of type.receivers) {
                         const selector = r.selector;
-                        if (selector.kind === 'internal-comment') {
+                        if (selector.kind === (internal ? 'internal-comment' : 'external-comment')) {
                             let hash = beginCell()
                                 .storeUint(0, 32)
                                 .storeBuffer(Buffer.from(selector.comment, 'utf8'))
@@ -163,7 +172,7 @@ export function writeRouter(type: TypeDescription, ctx: WriterContext) {
                             ctx.inIndent(() => {
 
                                 // Execute function
-                                ctx.append(`self~${ops.receiveText(type.name, hash)}();`);
+                                ctx.append(`self~${ops.receiveText(type.name, kind, hash)}();`);
 
                                 // Exit
                                 ctx.append('return (self, true);');
@@ -174,14 +183,14 @@ export function writeRouter(type: TypeDescription, ctx: WriterContext) {
                 }
 
                 // Comment fallback resolver
-                let fallback = type.receivers.find((v) => v.selector.kind === 'internal-comment-fallback');
+                let fallback = type.receivers.find((v) => v.selector.kind === (internal ? 'internal-comment-fallback' : 'external-comment-fallback'));
                 if (fallback) {
 
                     ctx.append(`if (slice_bits(in_msg) >= 32) {`);
                     ctx.inIndent(() => {
 
                         // Execute function
-                        ctx.append(`self~${ops.receiveAnyText(type.name)}(in_msg.skip_bits(32));`);
+                        ctx.append(`self~${ops.receiveAnyText(type.name, kind)}(in_msg.skip_bits(32));`);
 
                         // Exit
                         ctx.append('return (self, true);');
@@ -194,14 +203,14 @@ export function writeRouter(type: TypeDescription, ctx: WriterContext) {
         }
 
         // Fallback
-        let fallbackReceiver = type.receivers.find((v) => v.selector.kind === 'internal-fallback');
+        let fallbackReceiver = type.receivers.find((v) => v.selector.kind === (internal ? 'internal-fallback' : 'external-fallback'));
         if (fallbackReceiver) {
 
             ctx.append();
             ctx.append(`;; Receiver fallback`);
 
             // Execute function
-            ctx.append(`self~${ops.receiveAny(type.name)}(in_msg);`);
+            ctx.append(`self~${ops.receiveAny(type.name, kind)}(in_msg);`);
 
             ctx.append('return (self, true);');
         } else {
@@ -220,12 +229,12 @@ export function writeReceiver(self: TypeDescription, f: ReceiverDescription, ctx
     let selfUnpack = `var ${resolveFuncTypeUnpack(self, id('self'), ctx)} = ${id('self')};`;
 
     // Binary receiver
-    if (selector.kind === 'internal-binary') {
+    if (selector.kind === 'internal-binary' || selector.kind === 'external-binary') {
         let args = [
             selfType + ' ' + id('self'),
             resolveFuncType(selector.type, ctx) + ' ' + id(selector.name)
         ];
-        ctx.append(`((${selfType}), ()) ${ops.receiveType(self.name, selector.type)}(${args.join(', ')}) impure inline {`);
+        ctx.append(`((${selfType}), ()) ${ops.receiveType(self.name, selector.kind === 'internal-binary' ? 'internal' : 'external', selector.type)}(${args.join(', ')}) impure inline {`);
         ctx.inIndent(() => {
             ctx.append(selfUnpack);
             ctx.append(`var ${resolveFuncTypeUnpack(selector.type, id(selector.name), ctx)} = ${id(selector.name)};`);
@@ -244,8 +253,8 @@ export function writeReceiver(self: TypeDescription, f: ReceiverDescription, ctx
     }
 
     // Empty receiver
-    if (selector.kind === 'internal-empty') {
-        ctx.append(`((${selfType}), ()) ${ops.receiveEmpty(self.name)}(${(selfType + ' ' + id('self'))}) impure inline {`);
+    if (selector.kind === 'internal-empty' || selector.kind === 'external-empty') {
+        ctx.append(`((${selfType}), ()) ${ops.receiveEmpty(self.name, selector.kind === 'internal-empty' ? 'internal' : 'external')}(${(selfType + ' ' + id('self'))}) impure inline {`);
         ctx.inIndent(() => {
             ctx.append(selfUnpack);
 
@@ -263,14 +272,14 @@ export function writeReceiver(self: TypeDescription, f: ReceiverDescription, ctx
     }
 
     // Comment receiver
-    if (selector.kind === 'internal-comment') {
+    if (selector.kind === 'internal-comment' || selector.kind === 'external-comment') {
         let hash = beginCell()
             .storeUint(0, 32)
             .storeBuffer(Buffer.from(selector.comment, 'utf8'))
             .endCell()
             .hash()
             .toString('hex', 0, 64);
-        ctx.append(`(${selfType}, ()) ${ops.receiveText(self.name, hash)}(${(selfType + ' ' + id('self'))}) impure inline {`);
+        ctx.append(`(${selfType}, ()) ${ops.receiveText(self.name, selector.kind === 'internal-comment' ? 'internal' : 'external', hash)}(${(selfType + ' ' + id('self'))}) impure inline {`);
         ctx.inIndent(() => {
             ctx.append(selfUnpack);
 
@@ -289,8 +298,8 @@ export function writeReceiver(self: TypeDescription, f: ReceiverDescription, ctx
 
 
     // Fallback
-    if (selector.kind === 'internal-comment-fallback') {
-        ctx.append(`(${selfType}, ()) ${ops.receiveAnyText(self.name)}(${([selfType + ' ' + id('self'), 'slice ' + id(selector.name)]).join(', ')}) impure inline {`)
+    if (selector.kind === 'internal-comment-fallback' || selector.kind === 'external-comment-fallback') {
+        ctx.append(`(${selfType}, ()) ${ops.receiveAnyText(self.name, selector.kind === 'internal-comment-fallback' ? 'internal' : 'external')}(${([selfType + ' ' + id('self'), 'slice ' + id(selector.name)]).join(', ')}) impure inline {`)
         ctx.inIndent(() => {
             ctx.append(selfUnpack);
 
@@ -309,7 +318,7 @@ export function writeReceiver(self: TypeDescription, f: ReceiverDescription, ctx
 
     // Fallback
     if (selector.kind === 'internal-fallback') {
-        ctx.append(`(${selfType}, ()) ${ops.receiveAny(self.name)}(${selfType} ${id('self')}, slice ${id(selector.name)}) impure inline {`);
+        ctx.append(`(${selfType}, ()) ${ops.receiveAny(self.name, selector.kind === 'internal-fallback' ? 'internal' : 'external')}(${selfType} ${id('self')}, slice ${id(selector.name)}) impure inline {`);
         ctx.inIndent(() => {
             ctx.append(selfUnpack);