+++ src/grammar/ast.ts
@@ -261,7 +261,14 @@ export type ASTArgument = {
     ref: ASTRef
 }
 
-export type ASTFunctionAttribute = { type: 'public', ref: ASTRef } | { type: 'get', ref: ASTRef } | { type: 'mutates', ref: ASTRef } | { type: 'extends', ref: ASTRef } | { type: 'virtual', ref: ASTRef } | { type: 'overrides', ref: ASTRef };
+export type ASTFunctionAttribute =
+    | { type: 'public', ref: ASTRef }
+    | { type: 'get', ref: ASTRef }
+    | { type: 'mutates', ref: ASTRef }
+    | { type: 'extends', ref: ASTRef }
+    | { type: 'virtual', ref: ASTRef }
+    | { type: 'overrides', ref: ASTRef }
+    | { type: 'inline', ref: ASTRef };
 
 export type ASTFunction = {
     kind: 'def_function',
@@ -279,13 +286,19 @@ export type ASTReceive = {
     kind: 'def_receive',
     id: number,
     selector: {
-        kind: 'simple', arg: ASTArgument
+        kind: 'internal-simple', arg: ASTArgument
     } | {
-        kind: 'fallback'
+        kind: 'internal-fallback'
     } | {
-        kind: 'comment', comment: ASTString
+        kind: 'internal-comment', comment: ASTString
     } | {
         kind: 'bounce', arg: ASTArgument
+    } | {
+        kind: 'external-simple', arg: ASTArgument
+    } | {
+        kind: 'external-fallback'
+    } | {
+        kind: 'external-comment', comment: ASTString
     },
     statements: ASTStatement[],
     ref: ASTRef