+++ src/grammar/grammar.ts
@@ -160,6 +160,9 @@ semantics.addOperation<ASTFunctionAttribute>('resolve_attributes', {
     FunctionAttribute_overrides(arg0) {
         return { type: 'overrides', ref: createRef(this) };
     },
+    FunctionAttribute_inline(arg0) {
+        return { type: 'inline', ref: createRef(this) };
+    },
     FunctionAttribute_virtual(arg0) {
         return { type: 'virtual', ref: createRef(this) };
     },
@@ -306,7 +309,7 @@ semantics.addOperation<ASTNode>('resolve_declaration', {
     ReceiveFunction_simple(arg0, arg1, arg2, arg3, arg4, arg5, arg6) {
         return createNode({
             kind: 'def_receive',
-            selector: { kind: 'simple', arg: arg2.resolve_declaration() },
+            selector: { kind: 'internal-simple', arg: arg2.resolve_declaration() },
             statements: arg5.children.map((v: any) => v.resolve_statement()),
             ref: createRef(this)
         })
@@ -314,7 +317,7 @@ semantics.addOperation<ASTNode>('resolve_declaration', {
     ReceiveFunction_empty(arg0, arg1, arg2, arg3, arg4, arg5) {
         return createNode({
             kind: 'def_receive',
-            selector: { kind: 'fallback' },
+            selector: { kind: 'internal-fallback' },
             statements: arg4.children.map((v: any) => v.resolve_statement()),
             ref: createRef(this)
         })
@@ -322,7 +325,7 @@ semantics.addOperation<ASTNode>('resolve_declaration', {
     ReceiveFunction_comment(arg0, arg1, arg2, arg3, arg4, arg5, arg6) {
         return createNode({
             kind: 'def_receive',
-            selector: { kind: 'comment', comment: arg2.resolve_expression() },
+            selector: { kind: 'internal-comment', comment: arg2.resolve_expression() },
             statements: arg5.children.map((v: any) => v.resolve_statement()),
             ref: createRef(this)
         })
@@ -339,6 +342,22 @@ semantics.addOperation<ASTNode>('resolve_declaration', {
         return createNode({
             kind: 'def_receive',
             selector: { kind: 'bounce', arg: arg2.resolve_declaration() },
+            statements: arg5.children.map((v: any) => v.resolve_statement()),
+            ref: createRef(this)
+        })
+    },
+    ReceiveFunction_externalSimple(arg0, arg1, arg2, arg3, arg4, arg5, arg6) {
+        return createNode({
+            kind: 'def_receive',
+            selector: { kind: 'external-simple', arg: arg2.resolve_declaration() },
+            statements: arg5.children.map((v: any) => v.resolve_statement()),
+            ref: createRef(this)
+        })
+    },
+    ReceiveFunction_externalComment(arg0, arg1, arg2, arg3, arg4, arg5, arg6) {
+        return createNode({
+            kind: 'def_receive',
+            selector: { kind: 'external-comment', comment: arg2.resolve_expression() },
             statements: arg5.children.map((v: any) => v.resolve_statement()),
             ref: createRef(this)
         })